Add a script to run tpu checkpoint benchmark
